Refactor access building into separate files
